Create database and established connection in Jupyter Notebook
Using dbeaver I created the database with sqlite as the driver. I have also established the connection using the SQLConnection Class which I have copied from a previous week's assets file. Tested this connection with a rudimentary query of the SQLITE_SCHEMA table to see the tables and indexes that have been created in the ATP_test schema